Added support for project:issue-number queries
Added support for project specific issues (Chromium is the still default when the project name is omitted). b:v8:989, b:skia:822 and so on.
This is not optimized at all (it is called several times per query - creating an object every time - instead of being cached and re-used), but so was the previous code (though it was much simpler).